+++ action/action-dsl/src/test/resources/scalable.groovy
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-import static com.powsybl.action.util.Scalable.*
+import static com.powsybl.iidm.modification.scalable.Scalable.*
 
 action('actionScale') {
     modifications {
