+++ iidm/iidm-modification/src/main/java/com/powsybl/iidm/modification/scalable/LoadScalable.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.action.util;
+package com.powsybl.iidm.modification.scalable;
 
 import com.powsybl.iidm.network.*;
 import org.slf4j.Logger;
@@ -13,7 +13,7 @@ import org.slf4j.LoggerFactory;
 import java.util.List;
 import java.util.Objects;
 
-import static com.powsybl.action.util.Scalable.ScalingConvention.*;
+import static com.powsybl.iidm.modification.scalable.Scalable.ScalingConvention.*;
 
 /**
  * @author Ameni Walha <ameni.walha at rte-france.com>
