+++ iidm/iidm-modification/src/test/java/com/powsybl/iidm/modification/scalable/DanglingLineScalableTest.java
@@ -4,9 +4,9 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.action.util;
+package com.powsybl.iidm.modification.scalable;
 
-import com.powsybl.action.util.Scalable.ScalingConvention;
+import com.powsybl.iidm.modification.scalable.Scalable.ScalingConvention;
 import com.powsybl.commons.PowsyblException;
 import com.powsybl.iidm.network.DanglingLine;
 import com.powsybl.iidm.network.Injection;
@@ -16,9 +16,9 @@ import org.junit.Test;
 
 import java.util.List;
 
-import static com.powsybl.action.util.Scalable.ScalingConvention.GENERATOR;
-import static com.powsybl.action.util.Scalable.ScalingConvention.LOAD;
-import static com.powsybl.action.util.ScalableTestNetwork.createNetworkWithDanglingLine;
+import static com.powsybl.iidm.modification.scalable.Scalable.ScalingConvention.GENERATOR;
+import static com.powsybl.iidm.modification.scalable.Scalable.ScalingConvention.LOAD;
+import static com.powsybl.iidm.modification.scalable.ScalableTestNetwork.createNetworkWithDanglingLine;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertSame;
